Import story loading module from vite-app virtual module
