+++ packages/storybook-builder-vite/codegen-importfn-script.ts
@@ -1,9 +1,10 @@
 import * as path from 'path';
+import slash from 'slash';
 import { normalizePath } from 'vite';
 import { listStories } from './list-stories';
 
 import type { Options } from '@storybook/core-common';
-
+import type { ExtendedOptions } from './types';
 /**
  * This file is largely based on https://github.com/storybookjs/storybook/blob/d1195cbd0c61687f1720fefdb772e2f490a46584/lib/core-common/src/utils/to-importFn.ts
  */
@@ -46,4 +47,33 @@ export async function generateImportFnScriptCode(options: Options) {
 
   // We can then call toImportFn to create a function that can be used to load each story dynamically.
   return (await toImportFn(stories)).trim();
+}
+
+const absoluteFilesToImport = (files: string[], name: string) =>
+  files.map((el, i) => `import ${name ? `* as ${name}_${i} from ` : ''}'/@fs/${normalizePath(el)}'`).join('\n');
+
+export async function generateVirtualStoryEntryCode(options: ExtendedOptions) {
+  const storyEntries = await listStories(options);
+  const resolveMap = storyEntries.reduce<Record<string, string>>(
+    (prev, entry) => ({ ...prev, [entry]: entry.replace(slash(process.cwd()), '.') }),
+    {}
+  );
+  const modules = storyEntries.map((entry, i) => `${JSON.stringify(entry)}: story_${i}`).join(',');
+
+  return `
+    ${absoluteFilesToImport(storyEntries, 'story')}
+
+    function loadable(key) {
+      return {${modules}}[key];
+    }
+    
+    Object.assign(loadable, {
+      keys: () => (${JSON.stringify(Object.keys(resolveMap))}),
+      resolve: (key) => (${JSON.stringify(resolveMap)}[key])
+    });
+
+    export function configStories(configure) {
+      configure(loadable, { hot: import.meta.hot }, false);
+    }
+  `.trim();
 }
