+++ packages/storybook-builder-vite/input/iframe.html
@@ -32,6 +32,5 @@
     <div id="docs-root"></div>
 
     <script type="module" src="/virtual:/@storybook/builder-vite/vite-app.js"></script>
-    <script type="module" src="/virtual:/@storybook/builder-vite/storybook-stories.js"></script>
   </body>
 </html>
